Update `@typescript-eslint` package versions
The linter shows a warning on each run stating that the current version
of TypeScript is not supported. It looks like this:

```
=============

WARNING: You are currently running a version of TypeScript which is not officially supported by @typescript-es
lint/typescript-estree.

You may find that it works just fine, or you may not.

SUPPORTED TYPESCRIPT VERSIONS: >=3.3.1 <4.3.0

YOUR TYPESCRIPT VERSION: 4.3.5

Please only submit bug reports when using the officially supported version.

=============
```

The two `@typescript-eslint` packages have been updated to the latest
4.x versions. These version are compatible with TypeScript 4.3, so this
update resolves the warning.